IntelligentClassRoom/requirements.txt
Flask==3.0.0
Flask_Migrate==4.0.5
flask_sqlalchemy==3.1.1
pyserial==3.5
